Tags overhaul
Yet another overhaul of the tags system. I realised that tags could be a lot more useful than I was originally considering them, so it made sense to allow me to manage them in sanity itself. I also wanted to be able to apply as many colours to tags as I liked, so instead of using Tailwind colour classes, I've added a colour picker plugin to sanity and am now adding the colours as direct inline styles